[CLEANUP] Move `TestingMimeConvertableMailRole` to `Fixtures`
Fixes #1105
